Update README to reflect plugin option changes
