+++ dev-packages/e2e-tests/test-applications/solidstart/src/routes/index.tsx
@@ -15,6 +15,9 @@ export default function Home() {
           <A href="/server-error">Server error</A>
         </li>
         <li>
+          <A href="/error-boundary">Error Boundary</A>
+        </li>
+        <li>
           <A id="navLink" href="/users/5">
             User 5
           </A>
