+++ packages/solidstart/README.md
@@ -178,11 +178,9 @@ export default defineConfig({
   vite: {
     plugins: [
       sentrySolidStartVite({
-        sourceMapsUploadOptions: {
-          org: process.env.SENTRY_ORG,
-          project: process.env.SENTRY_PROJECT,
-          authToken: process.env.SENTRY_AUTH_TOKEN,
-        },
+        org: process.env.SENTRY_ORG,
+        project: process.env.SENTRY_PROJECT,
+        authToken: process.env.SENTRY_AUTH_TOKEN,
         debug: true,
       }),
     ],
